app/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
openai
pydantic
python-dotenv